# backend/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Agency Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /adsets:
    get:
      operationId: listAdsets
      tags: [adsets]
      summary: List ad sets
      parameters:
        - name: campaignId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of ad sets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Adset"
    post:
      operationId: createAdset
      tags: [adsets]
      summary: Create an ad set
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdsetInput"
      responses:
        "201":
          description: Ad set created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Adset"
  /adsets/{id}:
    get:
      operationId: getAdset
      tags: [adsets]
      summary: Get an ad set
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ad set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Adset"
    patch:
      operationId: updateAdset
      tags: [adsets]
      summary: Update an ad set
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdsetUpdate"
      responses:
        "200":
          description: Ad set updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Adset"
    delete:
      operationId: deleteAdset
      tags: [adsets]
      summary: Delete an ad set
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ad set deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # Ads
  /ads:
    get:
      operationId: listAds
      tags: [ads]
      summary: List ads
      parameters:
        - name: adsetId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of ads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ad"
    post:
      operationId: createAd
      tags: [ads]
      summary: Create an ad
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdInput"
      responses:
        "201":
          description: Ad created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ad"
  /ads/{id}:
    get:
      operationId: getAd
      tags: [ads]
      summary: Get an ad
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ad
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ad"
    patch:
      operationId: updateAd
      tags: [ads]
      summary: Update an ad
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdUpdate"
      responses:
        "200":
          description: Ad updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ad"
    delete:
      operationId: deleteAd
      tags: [ads]
      summary: Delete an ad
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ad deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # Insights
...
  /reports:
    get:
      operationId: listReports
      tags: [reports]
      summary: List reports
      responses:
        "200":
          description: List of reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"
    post:
      operationId: createReport
      tags: [reports]
      summary: Generate a report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportInput"
      responses:
        "201":
          description: Report created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
  /reports/{id}:
    get:
      operationId: getReport
      tags: [reports]
      summary: Get a report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
    delete:
      operationId: deleteReport
      tags: [reports]
      summary: Delete a report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Report deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # AI Insights
...
  /meta/campaigns/{clientId}:
    get:
      operationId: getMetaCampaigns
      tags: [meta]
      summary: Fetch campaigns from Meta Ads API
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Meta campaigns
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MetaCampaignsResponse"
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    # Auth
...
    Adset:
      type: object
      required: [id, name, campaignId, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        campaignId:
          type: integer
        metaAdsetId:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, paused, completed, draft]
        dailyBudget:
          type: ["number", "null"]
        targeting:
          type: ["string", "null"]
        optimizationGoal:
          type: ["string", "null"]
        spend:
          type: ["number", "null"]
        impressions:
          type: ["integer", "null"]
        clicks:
          type: ["integer", "null"]
        createdAt:
          type: string
    AdsetInput:
      type: object
      required: [name, campaignId]
      properties:
        name:
          type: string
        campaignId:
          type: integer
        dailyBudget:
          type: number
        targeting:
          type: string
        optimizationGoal:
          type: string
    AdsetUpdate:
      type: object
      properties:
        name:
          type: string
        status:
          type: string
          enum: [active, paused, completed, draft]
        dailyBudget:
          type: number
        targeting:
          type: string
        optimizationGoal:
          type: string

    # Ad
    Ad:
      type: object
      required: [id, name, adsetId, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        adsetId:
          type: integer
        metaAdId:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, paused, disapproved, draft]
        headline:
          type: ["string", "null"]
        body:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        destinationUrl:
          type: ["string", "null"]
        spend:
          type: ["number", "null"]
        impressions:
          type: ["integer", "null"]
        clicks:
          type: ["integer", "null"]
        ctr:
          type: ["number", "null"]
        cpc:
          type: ["number", "null"]
        createdAt:
          type: string
    AdInput:
      type: object
      required: [name, adsetId]
      properties:
        name:
          type: string
        adsetId:
          type: integer
        headline:
          type: string
        body:
          type: string
        imageUrl:
          type: string
        destinationUrl:
          type: string
    AdUpdate:
      type: object
      properties:
        name:
          type: string
        status:
          type: string
          enum: [active, paused, disapproved, draft]
        headline:
          type: string
        body:
          type: string
        imageUrl:
          type: string
        destinationUrl:
          type: string

    # Insight
...
    Report:
      type: object
      required: [id, title, type, status, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        type:
          type: string
          enum: [weekly, monthly, campaign, revenue, ai_summary]
        clientId:
          type: ["integer", "null"]
        clientName:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, generating, ready, failed]
        fileUrl:
          type: ["string", "null"]
        summary:
          type: ["string", "null"]
        dateFrom:
          type: ["string", "null"]
        dateTo:
          type: ["string", "null"]
        createdAt:
          type: string
    ReportInput:
      type: object
      required: [title, type]
      properties:
        title:
          type: string
        type:
          type: string
          enum: [weekly, monthly, campaign, revenue, ai_summary]
        clientId:
          type: integer
        dateFrom:
          type: string
        dateTo:
          type: string

    # AI Insight
...
    MetaCampaign:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        status:
          type: string
        objective:
          type: string
        daily_budget:
          type: string
        spend:
          type: string
    MetaCampaignsResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/MetaCampaign"
        error:
          type: object
          properties:
            message:
              type: string